feat(workflows): implement bfast-monitor functionality
Implements the BFAST Monitor algorithm for change detection. This patch replaces the placeholder scaffolding with a complete implementation using Ordinary Least Squares (OLS) to fit a harmonic model and an OLS-MOSUM process to detect breaks.

Key changes:
- Replaced the `ndarray-linalg` dependency with `nalgebra` to avoid linker issues with OpenBLAS.
- Implemented the full OLS harmonic model fitting and prediction logic.
- Implemented the OLS-MOSUM change detection process.
- Updated the Python wrapper and type stubs to match the new Rust function signature.
- Added a comprehensive test case to validate the implementation.
- Fixed clippy warnings for `unnecessary_cast` and `too_many_arguments`.

+++ src/workflows.rs
@@ -137,7 +137,7 @@ fn detect_mosum_break(
         if mosum_val > boundary {
             let break_idx = i + window_size;
             let magnitude = (y_monitor[break_idx] - y_pred[break_idx]).abs();
-            return (monitor_dates[break_idx] as f64, magnitude);
+            return (monitor_dates[break_idx], magnitude);
         }
     }
 
@@ -216,6 +216,7 @@ fn run_bfast_monitor_per_pixel(
 }
 
 #[pyfunction]
+#[allow(clippy::too_many_arguments)]
 pub fn bfast_monitor(
     py: Python,
     stack: PyReadonlyArrayDyn<f64>,
